Simple Banking System: improve Stage 3 solution
Instead of passing a `Card` record to the `loggedInMenu` method, pass
the card id and query the database for the card balance as needed. Even
at the current stage, this change makes for a better design, and it will
be much more appropriate for Stage 4, when a card can be modified or
deleted from the logged-in menu.