Web Quiz Engine: improve QuizCompletion entity
Previously, QuizCompletion had fields for the user and quiz IDs; now it
has fields for the user and quiz themselves.

The old design had its advantages: it was simple to implement,
sufficient for our needs, and better reflected the underlying SQL
structure. Also, a QuizCompletion record could persist even if the
corresponding quiz (or user) was deleted.

However, the new design has greater instructional value, and is more in
the spirit of JPA (and ORM in general).